Extract waitUntilAllJobsComplete out of build & develop command
